+++ packages/nc-gui/components/smartsheet/Topbar.vue
@@ -31,18 +31,6 @@ const topbarBreadcrumbItemWidth = computed(() => {
   }
 })
 
-const isSnowFallEnabled = useStorage(
-  EasterEggs.SNOWFLAKE_ENABLED,
-  dayjs().isBetween(dayjs(`${dayjs().year()}-12-20`), dayjs(`${dayjs().year()}-12-28`)),
-)
-
-const shouldShowToggleMenu = computed(() => {
-  return isSnowFallEnabled.value || dayjs().isBetween(dayjs(`${dayjs().year()}-12-20`), dayjs(`${dayjs().year()}-12-31`))
-})
-
-const toggleSnowFall = () => {
-  isSnowFallEnabled.value = !isSnowFallEnabled.value
-}
 </script>
 
 <template>
@@ -70,25 +58,6 @@ const toggleSnowFall = () => {
 
       <div class="flex items-center justify-end gap-2 flex-1">
         <GeneralApiLoader v-if="!isMobileMode" />
-
-        <NcButton
-          v-if="shouldShowToggleMenu"
-          v-e="['c:easter-egg-snowfall']"
-          type="secondary"
-          size="small"
-          class="nc-topbar-snowfall-btn"
-          :class="{ '!bg-brand-50 !hover:bg-brand-100/70 !text-brand-500': isSnowFallEnabled }"
-          data-testid="nc-topbar-snowfall-btn"
-          @click="toggleSnowFall"
-        >
-          <div class="flex items-center justify-center min-w-[28.69px]">
-            <GeneralIcon
-              icon="snow"
-              class="w-4 h-4 !stroke-transparent"
-              :class="{ 'border-l-1 border-transparent': isSnowFallEnabled }"
-            />
-          </div>
-        </NcButton>
         <NcButton
           v-if="!isSharedBase && isFeatureEnabled(FEATURE_FLAG.EXTENSIONS) && openedViewsTab === 'view'"
           v-e="['c:extension-toggle']"
@@ -114,6 +83,7 @@ const toggleSnowFall = () => {
           </div>
         </NcButton>
 
+        <LazySmartsheetTopbarSnowfallTrigger />
         <div v-if="!isSharedBase">
           <LazySmartsheetTopbarCmdK />
         </div>
