+++ packages/nc-gui/components/workspace/integrations/view.vue
@@ -58,6 +58,8 @@ onBeforeMount(() => {
         </h1>
       </div>
 
+      <LazySmartsheetTopbarSnowfallTrigger />
+
       <SmartsheetTopbarCmdK />
     </div>
     <NcTabs v-model:activeKey="activeViewTab">
